Day 1 progress for standalone analyzer application
Spent most of the last 12 hours hacking this together.  It's not very clean, but it appears to perform the basic functions I was aiming for without too much trouble; i.e. I can point to a solution file, browse the projects and their contents, and see the analyzer diagnostics for each of the C# source files.

Still TODO is:
1. Cleanup, and lots of it.
2. Making the code file summary view look better.
3. A different text marking method for the source view - hopefully one to provide tool tips for the highlighted issues.
4. Some sort of export functionality (CSV, Excel spreadsheet, etc...)?